requirements.txt
beautifulsoup4==4.12.2
numpy==1.25.2
pandas==2.0.3
Pillow==10.0.0
Pillow==10.1.0
Requests==2.31.0
streamlit==1.29.0
streamlit_extras==0.3.6
streamlit_option_menu==0.3.6
